Line, central symmetry, shape and color detection done

+++ Code/image.h
@@ -98,7 +98,7 @@ typedef struct {
 	int x;
 	int y;
 	int label;
-	int coef;
+	float coef;
 } Point;
